Fix the validate* function signatures in tests and documentation
The tests had wrong signatures in the validate* functions. All the
validate* functions take an `info` as a second argument. The
validate_<field_name> functions additionally take some keyword
arguments, thus receiving the whole mutation input.

The documentation was a bit incorrect as well, so fixed it.

Started calling the keyword arguments received by validate_<field_name>
as `input_args`, hopefully better describing that they mean the mutation
input arguments. And also because `input` is a built-in function in
Python so it's a bad name for an argument.