Add SingleValidationError class
This error class is for cases when the error contains exactly one error
detail. The example and test validation error classes in this project
are mostly this kind.

The ValidationError class is now more like an abstract base
class. It doesn't have the `meta`, `code` and `path` attributes any
more.